requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1
pydantic>=1.10,<3
requests>=2.31,<3
pillow>=10,<12
numpy>=1.26
pandas>=2.1
pyarrow>=14
faiss-cpu>=1.7.4
torch>=2.0
torchvision>=0.15
transformers>=4.40

